requirements.txt
discord.py
python-dotenv
langchain
langchain-community
langchain-openai
langchain-anthropic
langchain-experimental
chromadb
pymupdf